requirements.txt
h11~=0.14.0
Pillow~=10.1.0
pip~=22.3.1
distro~=1.8.0
wheel~=0.38.4
colorama~=0.4.6
anyio~=3.7.1
sniffio~=1.3.0
httpcore~=1.0.2
idna~=3.4
certifi~=2023.11.17
setuptools~=65.5.1
openai~=1.3.3
httpx~=0.25.1
pydantic~=2.5.1
python-dotenv~=1.0.0